gvt: check for COFFEELAKE when setting edid
When setting up the virtual display, the correct port were the virtual
display is connected has to be communicated to the user of GVT. The
default port is PORT_B, and .set_edid is called if the port differs.
Previously, a platform check for the coffeelake platform was missing,
which led to a missing .set_edid call to tell the user of GVT that the
correct port is PORT_D. This check is now added, so .set_edid is called
in the correct situations.

Signed-off-by: Stefan Hertrampf <stefan.hertrampf@cyberus-technology.de>
Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>