drm/i915/gvt: serialize context switch and workload completion
We have observed host crashes due to page faults for long running
guests with GVT enabled. The symptom is always the same:

<1>[ 2434.100389] BUG: unable to handle page fault for address: 00000000000263e0
<1>[ 2434.100398] #PF: supervisor write access in kernel mode
<1>[ 2434.100399] #PF: error_code(0x0002) - not-present page
<6>[ 2434.100400] PGD ffff888608e04000 0 P4D ffff888608e04000 0
<4>[ 2434.100402] Oops: 0002 [#1] SMP NOPTI
<4>[ 2434.100404] CPU: 2 PID: 2034 Comm: EMT-1 Tainted: P           O      5.4.68+ #1
<4>[ 2434.100405] Hardware name: LENOVO 20N2S00600/20N2S00600, BIOS N2IET00S (3.99 ) 07/30/2020
<4>[ 2434.100407] RIP: queued_spin_lock_slowpath (kernel/locking/qspinlock.c:471)
[...]
<4>[ 2434.100418] Call Trace:
<4>[ 2434.100420]  <IRQ>
<4>[ 2434.100423] _raw_spin_lock_irqsave (kernel/locking/spinlock.c:159)
<4>[ 2434.100425] __wake_up_common_lock (kernel/sched/wait.c:123 (discriminator 1))
<4>[ 2434.100429] shadow_context_status_change (drivers/gpu/drm/i915/gvt/scheduler.c:299)
<4>[ 2434.100431] notifier_call_chain (kernel/notifier.c:104)
<4>[ 2434.100433] atomic_notifier_call_chain (kernel/notifier.c:203)
<4>[ 2434.100435] process_csb (drivers/gpu/drm/i915/gt/intel_lrc.c:610 drivers/gpu/drm/i915/gt/intel_lrc.c:640 drivers/gpu/drm/i915/gt/intel_lrc.c:1590)
<4>[ 2434.100437] execlists_submission_tasklet (drivers/gpu/drm/i915/gt/intel_lrc.c:1637)
<4>[ 2434.100439] tasklet_action_common (./arch/x86/include/asm/bitops.h:75 ./include/asm-generic/bitops-instrumented.h:57 ./include/linux/interrupt.h:624 kernel/softirq.c:523)
<4>[ 2434.100441] __do_softirq (./arch/x86/include/asm/jump_label.h:25 ./include/linux/jump_label.h:200 ./include/trace/events/irq.h:142 kernel/softirq.c:293)
<4>[ 2434.100443] irq_exit (kernel/softirq.c:373 kernel/softirq.c:413)
<4>[ 2434.100445] do_IRQ (arch/x86/kernel/irq.c:267)
<4>[ 2434.100446] common_interrupt (arch/x86/entry/entry_64.S:890)
<4>[ 2434.100447]  </IRQ>

After some research, it seems this is the same issue that was already
reported to the mailing list in 2017 [1]. We see the workload being
completed (complete_current_workload) right before the
shadow_context_status_change notification comes in.

Reading the code it is clear that there is insufficient locking for
this case and we end up with a use-after-free in the notifier. Add a
workaround that serializes the particular case we are seeing.

This fix is very likely incomplete, because we do not understand how
we end up with the late notification to begin with.

v2: The first version of this patch protected almost the complete
destruction of the workload with the newly introduced spinlock. This
was ill-advised, as something in this code path seems to open
interrupts and we end up with hung tasks. Reduce the scope of the spin
lock to setting the current workload. This should be enough to avoid
the original problem.

[1] https://lists.freedesktop.org/archives/intel-gvt-dev/2017-February/000063.html

Inspired-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>