drm/i915/gvt: add implementation for SLVMGT
To enable graphics virtualization for SLVM, we used to carry a pretty
invasive patch to the GVT code to export a public interface for
hypervisor backends. The API was intended to be used by any hypervisor
backend. This turned out to be misguided, because:

a) The GVT interface turned out to be ugly and there was no obvious
   way to make it simpler.
b) Because we reached deeply into GVT, the patchset would often
   conflict during rebase.

This patch is our second generation of the SLVM graphics
virtualization patchset that tries to leave GVT alone and only exposes
a special purpose interface for SLVM.

Update: This patch includes the latest changes, necessary to support
kernels above 5.14

Update v2: We now also attach on systems where GVT is not available.
This allows SLVM to work on these systems as well.

Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>
Signed-off-by: Martin Messer <martin.messer@cyberus-technology.de>
Signed-off-by: Markus Partheymüller <markus.partheymueller@cyberus-technology.de>