platform/x86: think-lmi: expose 'type' attribute
think-lmi currently doesn't adhere to sysfs layout for
firmware-attributes. It doesn't expose a 'type' file. As we already
expose 'possible_values', we need to set the type to "enumeration".

This fixes the following fwupd error message:

   sudo fwupdtool security
[sudo] password for julian:
Loading                 [*************************              ]
12:27:57:0662 FuBiosSettings       KERNEL BUG: 'type' attribute not exported: (failed to load type: Failed to open file /sys/class/firmware-attributes/thinklmi/attributes/SecureBoot/type: No such file or directory)
Loading                 [-                                      ]

Not enough data was provided to make an HSI calculation.
  https://fwupd.github.io/hsi.html#not-enough-info
To ignore this warning, use --force

Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>