KVM: x86: Fix partially uninitialized integer in emulate_pop
Most code gives a pointer to an uninitialized unsigned long as dest in
emulate_pop. len is usually the word width of the guest.

If the guest runs in 16-bit or 32-bit modes, len will not cover the
whole unsigned long and we end up with unitialized data in dest.

Looking through the rest of the code, this seems harmless, but given
that none of this is performance critical, there should be no issue
with just always initializing the whole value.

Fix this by explicitly requiring a unsigned long pointer and
initializing it with zero in all cases.

Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>