WIP KVM: x86: Allow minimizing emulator for Linux/Windows guests
When we run Linux or Windows, only a fraction of the instruction
emulator is actually used. The rest is just a liability from a
security perspective. If we emulate RETF or SYSEXIT, for example,
something surely has gone wrong in the guest.

This patch introduces the config option KVM_MINIMAL_EMULATOR that cuts
down the emulator to instructions that we reasonably expect a 64-bit
Linux or Windows guest to use.

I've been a bit conservative and didn't exclude the arithmetic ops
that are implemented as fastops, but in theory most of that could be
removed as well.

Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>